Rename runtime_shared.js -> runtime_common.js. NFC
`shared can be misleading because its used when referring to shared
memory, etc.
