+++ src/runtime_common.js
